Creating this new file
Referencing the entire takeStockOfPhotos (currently named something else on here) made the program run every time making it so that it didn't really append os much as write a new file every time so that the comparisons made no sense. 

This separates out the helper methods so that they can be referenced without the main file creation program.